many changes
Move picture of me on the left and logos on the right.  Put my email
address under the picture.

Replace my email address with the course title in the footline.

Use different colors for ticks and crosses.

Number blocks.

Author name in small caps.

Add the QR code for the arXiv preprint.